fix:fix see more page
