+++ clients/src/pages/AuthenticatedApp.jsx
@@ -7,7 +7,7 @@ import Register from "./Register.jsx";
 import { ThemeProvider } from "@mui/material";
 import { theme } from "../theme.js";
 import { Recruiter } from "./Recruiter.jsx";
-import { SeeMoreJobDetails } from "../components/Professional/SeeMoreJobDetails.jsx";
+import { SeeMorePage } from "./SeeMorePage.jsx";
 
 function AuthenticationApp() {
   return (
@@ -16,7 +16,7 @@ function AuthenticationApp() {
         <Route path="/" element={<HomePage />}></Route>
         <Route path="/login" element={<Login />}></Route>
         <Route path="/findjobs" element={<FindJobs />}></Route>
-        <Route path="/findjobs/:jobId" element={<SeeMoreJobDetails />}></Route>
+        <Route path="/findjobs/:jobId" element={<SeeMorePage />}></Route>
         {/* -------------------------mock recruiter createpost------------------ */}
         <Route path="/recruiter" element={<Recruiter />}></Route>
         {/* -------------------------mock recruiter createpost------------------ */}
